refactor(benchmark): split admission acceptance policy
